[build] Make v8_enable_cluster_build
New argument for the gn build can be set to
true to enable a faster build.  This compiles
.cc files together that have a similar compile
time and are in the same directory.

For example, profiler-cluster.cc replaces 4
.cc files, each of which takes over 12s to compile.
The 4 files together take 13.7s to compile.

This feature is off by default, but it is intended
to be usable by default and should cause no noticable
ill effects even for iterative changes to a single
.cc file.

In this version of the change there are many files
excluded from the clusters, rather than fixing their
top level name clashes.

+++ src/objects/elements.cc
@@ -1447,8 +1447,8 @@ class ElementsAccessorBase : public InternalElementsAccessor {
         // large-object space which doesn't free memory on shrinking the list.
         // Hence we try to estimate the final size for holey backing stores more
         // precisely here.
-        initial_list_length =
-            Subclass::NumberOfElementsImpl(isolate, *object, *backing_store);
+        initial_list_length = static_cast<uint32_t>(
+            Subclass::NumberOfElementsImpl(isolate, *object, *backing_store));
         initial_list_length += nof_property_keys;
       }
       DCHECK_LE(initial_list_length, std::numeric_limits<int>::max());
